+++ src/lib.rs
@@ -49,11 +49,9 @@ numeric primitives are implemented. That is, you can use [`From`] or
 ## Quick example
 
 ```rust
-use fixed::frac;
-use fixed::FixedI32;
-
 // 20 integer bits, 12 fractional bits
-type I20F12 = FixedI32<frac::U12>;
+use fixed::types::I20F12;
+
 // 19/3 = 6 1/3
 let six_and_third = I20F12::from_int(19).unwrap() / 3;
 // four decimal digits for 12 binary digits
@@ -61,6 +59,12 @@ assert_eq!(six_and_third.to_string(), "6.3333");
 // convert to i32, taking the ceil
 assert_eq!(six_and_third.to_int_ceil(), 7);
 ```
+
+The type [`I20F12`] is a 32-bit fixed-point signed number with 20
+integer bits and 12 fractional bits. It is an alias to
+[`FixedI32<frac::U12>`][`FixedI32`]. The unsigned counterpart would be
+[`U20F12`]. Aliases are provided for all combinations of integer and
+fractional bits adding up to a total of eight, 16, 32, 64 or 128 bits.
 
 ## Using the *fixed* crate
 
@@ -132,7 +136,9 @@ additional terms or conditions.
 [`FixedU64`]: struct.FixedU64.html
 [`FixedU8`]: struct.FixedU8.html
 [`From`]: https://doc.rust-lang.org/nightly/std/convert/trait.From.html
+[`I20F12`]: types/type.I20F12.html
 [`Into`]: https://doc.rust-lang.org/nightly/std/convert/trait.Into.html
+[`U20F12`]: types/type.U20F12.html
 [`f16`]: https://docs.rs/half/^1/half/struct.f16.html
 [const generics]: https://github.com/rust-lang/rust/issues/44580
 */